third_party/WebKit/LayoutTests/editing/execCommand/insert_paragraph/root-htmlelement-editable.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => {
assert_selection(
  'foobar|baz',
  selection => {
    const doc = selection.document;
    doc.documentElement.contentEditable = true;
    doc.execCommand("InsertParagraph");
  },
  'foobar<div>|baz</div>');
}, 'InsertParagraph BODY when parent HTML element is contenteditable');
</script>
